Simple blog is complete.Added some auth, error 404.

+++ app/Http/Controllers/PostController.php
@@ -6,11 +6,18 @@ use App\Http\Requests\PostRequest;
 use App\Post;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+
+    public function __construct()
+{
+    $this->middleware('auth')->except('index','show');
+}
+
     /**
      * Display a listing of the resource.
      *
@@ -59,10 +66,10 @@ class PostController extends Controller
     public function store(PostRequest $request)
     {
         $post = new Post();
-        $post->title=$request->title;
-        $post->author_id=rand(1,19);
+        $post->title = $request->title;
+        $post->author_id = Auth::user()->id;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
-        $post->description=$request->description;
+        $post->description = $request->description;
 
         if($request->file('img')){
             $path=Storage::putFile('public',$request->file('img'));
@@ -83,6 +90,9 @@ class PostController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+        if(!$post){
+            return redirect()->route('posts.index')->withErrors('Такой страницы не найдено!');
+        }
         return view('posts.show',compact('post'));
     }
 
@@ -95,6 +105,12 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(!$post){
+            return redirect()->route('posts.index')->withErrors('Такой страницы не найдено!');
+        }
+        if($post->author->id !== Auth::user()->id){
+            return redirect()->route('posts.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         return view('posts.edit',compact('post'));
     }
 
@@ -108,6 +124,12 @@ class PostController extends Controller
     public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+        if(!$post){
+            return redirect()->route('posts.index')->withErrors('Такой страницы не найдено!');
+        }
+        if($post->author->id !== Auth::user()->id){
+            return redirect()->route('posts.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         $post->title=$request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->description=$request->description;
@@ -130,6 +152,12 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(!$post){
+            return redirect()->route('posts.index')->withErrors('Такой страницы не найдено!');
+        }
+        if($post->author->id !== Auth::user()->id){
+            return redirect()->route('posts.index')->withErrors('Вы не можете удалить данный пост');
+        }
         $post->delete();
         return redirect()->route('posts.index')->with('success','Пост успешно удален!');
     }
